Provide an implementation for QPlatformSystemTrayIcon
The idea is to force all QSystemTrayIcon to use our status notifiers
as we don't want to provide an xembed based system tray in the next
iteration of the Plasma desktop shell anymore.

The KDEPlatformSystemTrayIcon uses a KStatusNotifierItem to implement
the system tray icon. Unfortunately a complete wrapping is not yet
possible as we cannot create a menu. We do not want to provide a
QPlatformMenu in our PlatformTheme and thus the menu used by
QSystemTrayIcon does not have a QPlatformMenu.

This is adressed in Qt 5.3 which extends the QPA API.