Import KDirSelectDialog + use it if QFileDialog requests dirs only
This imports KDirSelectDialog from kde4support into
frameworksintegration (+ KFileTreeView) and uses the dir select dialog
in case the user wants to select a directory.

To not disturb the current code too much, I've created a new class
KDEPlatformFileDialogBase, made KDEPlatformFileDialog derive from that
as well as KDirSelectDialog, and then KDEPlatformFileDialogHelper stores
the KDEPlatformFileDialogBase pointer, which is either the file selector
or dialog selector, but with the same API. The API is just added to
KDirSelectDialog and it calls its own functions internally.

REVIEW: 117435