fetch some defaults from a Plasma L&F package
there will be support for some kind of "mega theme" that besides providing files for splashscreen, lockscreen etc, will define what defaults to use among icons, colors, fonts etc.

This is a first start of it for reading the defaults. The theme file is stored in a plasma package, but the patch resolves the correct path by hand, not requiring linking to libplasma.