Fix broken unittest for QFileDialog::setViewMode.
Commit 0b8ab82e7fa9ea0948a841ebf8523b62da made the view mode a choice of
the user (as it always was in KDE code) rather than a choice of the
developer (as QFileDialog::setViewMode allows).

Typical Qt/KDE difference: Qt empowers the developer, KDE empowers the user.
I agree with Lukas that the latter makes more sense in KDE code.

=> I rewrote the unittest to check that the user's changes are restored
on the next file dialog.

REVIEW: 121747