Fix unit test failure on machines with an empty ~/.qttest.
To execute the KdePlatformTheme and the KStyle unit tests, a local
kdeglobals test file is required inside the Qt test folders.  If the Qt
test folders don't exist when the test executable starts, this copy will
silently fail, causing further failures.

Now, attempt to pre-create this folder before the copy.  Also abort
the test if the folder creation or the file copy fail, to help diagnose
this issue in the future.

REVIEW: 121908