Fix native file dialogs from widgets QFileDialog
There were two issues:
 * File dialogs opened with exec() and without parent were
   opened, but any user-interaction was blocked in a way that
   no file could be selected nor the dialog closed.

 * File dialogs opened with open() or show() with parent were
   not opened at all.

The first issue was caused by first calling show() and then exec()
on the native dialog.
The second one simply because in the case of dialogs with parent
show() wasn't called.

This fixes it that the show() is always called and hide() is called
before exec().

This also adds unittests for file dialogs.

REVIEW: 123335