Depend on Breeze to set the default style name
Summary:
Instead of hard coding the name to Breeze, we now find Breeze as a
hard dependency and use the CMake variable of the name to generate
the name.

This allows us to have a proper dependency set and we can ensure that
Breeze is installed when using Plasma.

Reviewers: #plasma

Subscribers: plasma-devel

Projects: #plasma

Differential Revision: https://phabricator.kde.org/D1397