Auto-test to verify that the empty defaults are loaded correctly in KHintsSettings
Summary:
This change introduces a new auto-test for KHintsSettings. The idea is to
verify that defaults are loaded correctly if there is no kdeglobals at
all.

To make it possible to have such a test the ctor of KHintSettings is
adjusted to allow an optional KSharedConfigPtr. Thus the test can add
a dummy config.

Reviewers: #plasma

Subscribers: plasma-devel

Projects: #plasma

Differential Revision: https://phabricator.kde.org/D1455