Set color scheme X property in the X11Integration
Summary:
So for the _KDE_NET_WM_COLOR_SCHEME property was set in KStyle. This has
a few disadvantages:
* it only works for QWidgets
* it requires a style inheriting from KStyle

This change brings the behavior into the plasma-integration QPT. It's an
integration feature for our platform theme as it's a hint for KWin.

The existing event filter in the X11Integration is used to monitor for
all PlatformSurface events and sets the required property on the newly
created X11 window.

In addition it monitors for the ApplicationPaletteChange to update the
property as needed on all toplevel windows.

Test Plan:
Removed code from kstyle and used the KColorSchemeManager test
application.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D3044