Install desktop file name on X11 windows
Summary:
KWindowSystem provides a way to set the desktop file name on X11
windows. As Qt 5.7 also gives us the desktop file name we can export it
to the windows. This brings the advantage that interested parties
(window manager, task manager) can map the window to the desktop file.
Thus take the icon from the desktop file in high quality instead of a
lower quality one exported to the window.

Also it brings a nice advantage for things like kcmshell5. Instead of
having it mapped to kcmshell5 it is mapped to the actual open config
module.

Reviewers: #plasma, hein

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D3175