[KHintsSettings] Emit QGuiApplication::paletteChanged when run as QApplication
QGuiApplication::setPalette and QApplication::setPalette are two distinct static functions and
when the paletteChanged signal was introduced in QGuiApplication the latter was never adjusted.

SystemPalette connects to this signal to avoid installing an eventFilter on qApp for ApplicationPaletteChanged.
This causes e.g. QML-based System Settings modules to not update when changing color schemes but only the
widget-based application around it.

CHANGELOG: QtQuick-based interfaces will now more reliably adapt to color scheme changes at runtime

Differential Revision: https://phabricator.kde.org/D4492