Use QQuickStyle to set the QQC2 style
Use QQuickStyle::setStyle instead of an environment variable to set the QQC2 desktop style. This prevent the style from being inherited by
plasmashell (or krunner) child processes, which makes QGuiApplications crash. The QQC1 style still needs to be set via an environment variable, so
we explicitely unset it for non-QApplications to prevent them from crashing.

Differential Revision: https://phabricator.kde.org/D7953
BUG: 384481