Port plasma-integration to explicit wayland AppMenu protocol
Summary:
Also fixed to handle the case of surfaces not having the lifespan of the
window, which the old code didn't handle.

I don't think it's worth trying to handle the case of different versions
of kwin and plasma-integration.

Test Plan: Made relevant kwin change and it gets communicated fine

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D9248