Disable KDE ServerDecoration on Qt > 5.12
Summary:
XdgDecoration is built into QtWayland 5.12. Running two interfaces to do
the same thing would be messy.

We cannot just remove this code as Plasma 5.15 will depend on Qt5.11.

Detection of the ShellSurfaceInterface was used to test if we have run
shellSurfaceCreated which we still need for other functions, so this check
is replaced with a new explicit property.

Test Plan:
Got the correct decorations still.
I haven't tested on Qt5.11. I would appreciate someone checking that
things still work the same.

Reviewers: #kwin, #plasma, graesslin

Reviewed By: #kwin, #plasma, graesslin

Subscribers: graesslin, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D17758