Don't set a new palette if the application has a custom color scheme set
Summary:
If an application has explicitly set a custom color scheme we shouldn't change
the palette when the global color scheme changes because it explicitly wants to
use its custom one. For the same reason we already don't update the titlebar
color to the global one if this property is set.

Test Plan:
Change colorscheme in kate or kcolorschemedemo and change  the global
one in the colors kcm.

Reviewers: #plasma, ngraham

Reviewed By: ngraham

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25915