Resize file dialog window before size restore
Summary:
KDEPlatformDialogHelper explicitly resizes the window contents to the
window as part of a bug fix. This means we lose the common path where
the window resizes to the content's size hints that happens later.

If we resize the window to the size hint in advance we keep our size
hint.

(alternatively it seems to be me we can remove that old hack...but it
needs testing by people with Qt5.12)

Test Plan:
KWrite -> open

Size is 900x600, the size hint
not 640x480 the QPA default

Reviewers: #plasma, ngraham

Reviewed By: ngraham

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25986