Fix build
Summary:
Though for some reason it's not failing on the CI, plasma-integration is failing to build
for me with the following error:

```
/home/nate/kde/src/plasma-integration/src/platformtheme/kdeplatformfiledialoghelper.cpp:342:78: error: cannot bind non-const lvalue reference of type ‘QStringList&’ to an rvalue of type ‘QStringList’
  342 |         dialog->m_fileWidget->setSupportedSchemes(options()->supportedSchemes());
      |                                                   ~~~~~~~~~~~~~~~~~~~~~~~~~~~^~
```

`options()->supportedSchemes()` returns a const reference but `dialog->m_fileWidget->setSupportedSchemes`
wants a non-const reference, so this patch has it make a local copy to keep it happy. This
is kind of weird since the code here hasn't been touched in years and
`options()->supportedSchemes() has returned a const reference for a long time too, yet
this code only now started failing for me yesterday.

Test Plan: It builds for me now

Reviewers: davidre, #plasma, apol

Reviewed By: apol

Subscribers: apol, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28724