Raise size of default fixed-width font from 9 to 10pt
Summary:
See discussion in T12744. The fixed width font being smaller than others seems fairly
arbitrary, and we have gotten some user complaints about it being too small. And the devs
of our apps that use a lot of fixed-width fonts--Konsole and KTextEditor--are on board
with this change.

Reviewers: #vdg, #plasma, #ktexteditor, #kate, #konsole, ndavis

Reviewed By: #vdg, ndavis

Subscribers: ndavis, plasma-devel

Tags: #plasma

Maniphest Tasks: T12744

Differential Revision: https://phabricator.kde.org/D28841