Use KDE shortcuts for "move to trash" action coming from QStandardKey
QKeySequence::Delete is the Qt "Move to trash" action, which, by
default, includes Ctrl+D as a shortcut. This is something we don't want
in KDE software, so we need to intercept this action in our integration
plugin and give it the standard KDE shortcuts instead, as we do for
other actions.

BUG: 426573
FIXED-IN: 5.20


(cherry picked from commit 15c7926377065da0fea111be303816dca91d087a)