kdirselectdialog: when creating a new dir that already exists, select it
If the user tries to create a new dir and a dir with that name already
exists, select the existing dir in the tree view.

Don't use KIO::encodeFileName as it replaces '/' with the unicode character
fraction slash, which breaks creating nested dirs.

Drive-by change: port to KIO::setDetails(StatDetails).

Feature: 424082