Stop advertising titlebar "What's This?" help button
The "What's This?" feature has been transformed into an expandable
tooltip for KDE apps in frameworks/kxmlgui!45 (merged).

Now the text is shown by pressing the Shift key while one of these
expandable tooltips is visible. As a result, there is no longer a need
for a second redundant method of invoking the feature, so let's hide the
button to do so from dialog windows by setting
Qt::AA_DisableWindowContextHelpButton to true.