Draft: Add support for QPlatformTheme::SetFocusOnTouchRelease
Allow to set focus-on-release mode by setting env var
PLASMA_INTEGRATION_FOCUS_ON_RELEASE=1.

This can be useful for Plasma Mobile to fix pressing a button
in a dialog when a text field has a focus and thus a virtual
keyboard is visible, which then hides on press, the dialog
resizes and the button can't be pressed.