Revert "extend kio with portal-based open-with implementation"
This reverts commit 8c72d54596afe8e97c7bb96cccb324f24f67b7b0.
This reverts commit ae36cf167089b0bf42c29d501e9009b51980b4ec.

This work to unify on a single "Open With" dialog implementation is
important for the long-term, but it's not ready to be rolled out yet
due to missing features in the portal-based dialog that some users
currently rely on ("open with arbitrary binary" and "launch the chosen
app in a terminal window"). Because the roll-out happened for Plasma
5.27 LTS, the possibility of adding those features soon is unlikely
because the next Plasma feature release is 6.0 which is 8+ months way.

Let's revert this work for now to un-break users and stabilize the last
Plasma 5 LTS version, and then we can work on adding the missing
features for Plasma 6.0 and try again then.

BUG: 460741
CCBUG: 460985
FIXED-IN: 5.27.3