Bring back dbus menu functionality using Qt implementation
Drops the forked Qt code which didn't compile anymore since the symbols are no
longer available. Because of this inherit from QGenericUnixTheme in order to
instantiate a QDBusMenuBar via its createPlatformMenuBar implementation. We
have to  wrap this so we can register it properly on X11 and Wayland. Hopefully
Qt will do this itself in the future.
This also removes all the code handling "global menu bars". If the parent was
nullptr upstream QDBusMenuBar will not even register the menu bar via DBus.
Furthermore during testing the app supposedly broken by this - Qt Designer -
was found to actually have a parent for its menu bar.