qt6: install i18n after sources
curiously enough when doing it the other way around the library paths
on debians get bricked consistently. it's unclear to me why that happens
but it's reproducible in qt6-only sources as well so I suspect there is
some cache pollution going on. in any case we don't need this to run
earlier nor do we need it in its own if condition nor do we need to make
the ki18n_install call conditional (it is internally conditional
already)