Don't manually find the value of MainToolbarIcons
We can ask KIconLoader to do this for us, which lets us avoid manually
looking in the kdegglobals file and specifying a fallback size.
KIconLoader can do all of that internally.

This supports https://invent.kde.org/plasma/plasma-desktop/-/issues/58,
but would make sense to do even if that work ends up abandoned or
reverted.