Fix running autotests
There was no test target with the mixed Qt build, because creating the
BUILD_TESTING option and calling enable_testing is usually
KDECMakeSettings jobs, which isn't called in the top-level
CMakeLists.txt anymore.
Furthermore the version suffix was missing in the add_test call,
causing ctest failing to run the test executable.

Fixes #4.