Don't require breeze for building
It is only used the read the name of the style "breeze". We are
happy to hardcode the  other style names like "oxygen" as well.
It also simplifies the CMake situation for breeze between Qt5
and Qt6 versions.
I expect distros will not stop shipping breeze because we removed
the hard dependency here.