Implement Font dialogs
QML FontDialog first queries the platform theme for an implementation. Since we don't provide one
we get the default one, which is less than stellar.

In Qt5 we would get a widgets-based one that works fine.

We can get the Qt5 behavior back by asking for a QFontDialog (and passing DontUseNativeDialog to avoid endless recursion)

BUG: 476364