kdirselectdialog: correct initial focus
It begins with focus on the top-left-most widget, which normally is
good. However here that means the Places panel, which is not what you
want; you can't start navigating the tree view immediately, and if you
pres the Enter key, the top-most Places item gets returned instead of
the item visibly selected in the tree view.

Instead, start with focus on the main view, which corrects both issues.

BUG: 505046
FIXED-IN: 6.3.6