autotest: Provide a contrast value
Otherwise we will take the one of the active color scheme.
With these changes
palette->color(states[i], QPalette::Midlight).red() is 203 compared
to the expected 202 which seems was the intention but one off.
Same for QPalette::Dark with 84 and 85.
I suspect rounding issues in the color calculations which might
have changed or simple typo.