Simplify palette change handling
The current code does several things to change the palette

- It calls QApplication::setPalette
- It emits qApp->paletteChanged
- It sends a ApplicationPaletteChange event
- It calls QWindowSystemInterface::handleThemeChange

The upstream platform themes don't do most of that, only QWindowSystemInterface::handleThemeChange().

That seems to be entirely enough for Qt to pick up the change. It will re-read the palette automatically afterwards.

Not only does this massively simplify the code, it also fixes an issue with the colorScheme() function.

Currently when an application connects to paletteChanged() that slot will trigger before
Qt internally has fully processed the colorScheme change, because that internally relies on the ApplicationPaletteChange
event. By actively emitting paletteChanged earlier the slot will get the old value