Drop separate QStringList::contains check before QStringList::removeAll
removeAll will loop over whole list as well and just quit if nothing is
to be done.
removeAll might be a tad slower, but then one would assume that users
usually use the same small set of custom choices, so the current one
might be in the search/filter history already.